화면 post, 즐겨찾기 전환
