+++ app/src/main/java/com/im_geokjeong/ui/favorite/FavoriteFragment.kt
@@ -1,0 +1,35 @@
+package com.im_geokjeong.ui.favorite
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.core.os.bundleOf
+import androidx.fragment.app.Fragment
+import androidx.fragment.app.viewModels
+import androidx.lifecycle.lifecycleScope
+import androidx.navigation.fragment.findNavController
+import com.google.android.material.dialog.MaterialAlertDialogBuilder
+import com.im_geokjeong.R
+import com.im_geokjeong.model.PostPerson
+import com.im_geokjeong.ui.common.ViewModelFactory
+import com.im_geokjeong.ui.rentalpost.PostViewModel
+import kotlinx.coroutines.launch
+
+class FavoriteFragment : Fragment() {
+
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        return inflater.inflate(R.layout.fragment_favorite, container, false)
+    }
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+
+
+    }
+}
