+++ app/src/main/java/com/im_geokjeong/ui/rentalpost/PostActivity.kt
@@ -1,38 +1,23 @@
 package com.im_geokjeong.ui.rentalpost
 
+import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.core.os.bundleOf
-import androidx.fragment.app.Fragment
-import androidx.fragment.app.viewModels
+import androidx.activity.viewModels
 import androidx.lifecycle.lifecycleScope
-import androidx.navigation.fragment.findNavController
 import com.google.android.material.dialog.MaterialAlertDialogBuilder
 import com.im_geokjeong.R
 import com.im_geokjeong.model.PostPerson
 import com.im_geokjeong.ui.common.ViewModelFactory
-import kotlinx.android.synthetic.main.fragment_post.*
+import kotlinx.android.synthetic.main.activity_post.*
 import kotlinx.coroutines.launch
 
-class PostFragment : Fragment() {
-
-    private val viewModel: PostViewModel by viewModels { ViewModelFactory(requireContext()) }
+class PostActivity : AppCompatActivity() {
+    private val viewModel: PostViewModel by viewModels { ViewModelFactory(this) }
     private lateinit var person: PostPerson
 
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_post, container, false)
-    }
-
-    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
-        super.onViewCreated(view, savedInstanceState)
-
+    override fun onCreate(savedInstanceState: Bundle?) {
+        super.onCreate(savedInstanceState)
+        setContentView(R.layout.activity_post)
         postComplete.setOnClickListener {
             person = PostPerson(
                 title = postEditTitle.text.toString(),
@@ -45,7 +30,6 @@ class PostFragment : Fragment() {
             )
             lifecycleScope.launch {
                 viewModel.uploadPost(person)
-                finishPost()
             }
             //등록 완료되고 어떻게 할지?
         }
@@ -57,7 +41,7 @@ class PostFragment : Fragment() {
 
     private fun postClicked() {
         val items = arrayOf("Y", "N")
-        MaterialAlertDialogBuilder(requireContext())
+        MaterialAlertDialogBuilder(this)
             .setItems(items) { dialog, which ->
                 postEditProof.text = items[which]
             }
@@ -71,10 +55,4 @@ class PostFragment : Fragment() {
         }
     }
 
-    private fun finishPost() {
-        findNavController().navigate(
-            R.id.action_navigation_add_post_to_navigation_office, bundleOf(
-            )
-        )
-    }
 }
