+++ app/src/main/res/drawable/ic_plus.xml
@@ -1,0 +1,9 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="24"
+    android:viewportHeight="24">
+  <path
+      android:pathData="M10,18C10,19.1046 10.8954,20 12,20C13.1046,20 14,19.1046 14,18V14H18C19.1046,14 20,13.1046 20,12C20,10.8954 19.1046,10 18,10H14V6C14,4.8954 13.1046,4 12,4C10.8954,4 10,4.8954 10,6V10L6,10C4.8954,10 4,10.8954 4,12C4,13.1046 4.8954,14 6,14L10,14V18Z"
+      android:fillColor="#FFFFFF"/>
+</vector>
