+++ build.gradle
@@ -8,6 +8,7 @@ ext {
     navVersion = '2.3.5'
     glideVersion = '4.12.0'
     retrofitVersion = '2.9.0'
+    roomVersion = '2.4.1'
 }
 task clean(type: Delete) {
     delete rootProject.buildDir
